requirements.txt
adal==1.2.7
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.6.2.post1
applicationinsights==0.11.10
argcomplete==3.3.0
attrs==24.2.0
azure-ai-formrecognizer==3.3.3
azure-ai-ml==1.16.1
azure-appconfiguration==1.7.1
azure-batch==14.2.0
azure-cli==2.65.0
azure-cli-core==2.65.0
azure-cli-telemetry==1.1.0
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==3.2.0
azure-data-tables==12.4.0
azure-datalake-store==0.0.53
azure-graphrbac==0.60.0
azure-identity==1.19.0
azure-keyvault-administration==4.4.0b2
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.9.0b3
azure-keyvault-secrets==4.7.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-appconfiguration==3.0.0
azure-mgmt-appcontainers==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-batch==17.3.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-compute==33.0.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-containerservice==32.0.0
azure-mgmt-core==1.5.0
azure-mgmt-cosmosdb==9.6.0
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datamigration==10.0.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==10.2.0b2
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0b3
azure-mgmt-imagebuilder==1.3.0
azure-mgmt-iotcentral==10.0.0b2
azure-mgmt-iothub==3.0.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-managementgroups==1.0.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==5.0.1
azure-mgmt-msi==7.0.0
azure-mgmt-netapp==10.1.0
azure-mgmt-network==28.0.0
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.2.0b18
azure-mgmt-recoveryservices==3.0.0
azure-mgmt-recoveryservicesbackup==9.1.0
azure-mgmt-redhatopenshift==1.5.0
azure-mgmt-redis==14.4.0
azure-mgmt-resource==23.1.1
azure-mgmt-search==9.1.0
azure-mgmt-security==6.0.0
azure-mgmt-servicebus==8.2.0
azure-mgmt-servicefabric==2.1.0
azure-mgmt-servicefabricmanagedclusters==2.0.0b6
azure-mgmt-servicelinker==1.2.0b2
azure-mgmt-signalr==2.0.0b1
azure-mgmt-sql==4.0.0b19
azure-mgmt-sqlvirtualmachine==1.0.0b5
azure-mgmt-storage==21.2.0
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.2.0
azure-monitor-opentelemetry-exporter==1.0.0b31
azure-monitor-query==1.2.0
azure-multiapi-storage==1.3.0
azure-search-documents==11.5.2
azure-storage-blob==12.23.1
azure-storage-common==1.4.2
azure-storage-file-datalake==12.17.0
azure-storage-file-share==12.19.0
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.19.0
azure-synapse-managedprivateendpoints==0.4.0
azure-synapse-spark==0.2.0
azureml-core==1.58.0
azureml-dataprep==5.1.6
azureml-dataprep-native==41.0.0
azureml-dataprep-rslex==2.22.4
azureml-fsspec==1.3.1
azureml-rag==0.2.36
backoff==2.2.1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.2.0
beautifulsoup4==4.11.2
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
contextlib2==21.6.0
cryptography==43.0.3
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
emoji==2.14.0
fabric==3.2.2
fastapi==0.115.4
filelock==3.16.1
filetype==1.2.0
fixedint==0.1.6
Flask==3.0.3
Flask-Cors==4.0.2
flask-restx==1.3.0
frozenlist==1.5.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.22.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
invoke==2.2.0
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
javaproperties==0.5.2
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
jsondiff==2.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==24.3.1
knack==0.11.0
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.139
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.23.1
microsoft-security-utilities-secret-masker==1.0.0b3
mmh3==5.0.1
more-itertools==10.5.0
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.1.0
mypy-extensions==1.0.0
ndg-httpsclient==0.5.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
openai==1.44.1
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.11
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
pathspec==0.12.1
pillow==10.4.0
pkginfo==1.11.2
portalocker==2.10.1
promptflow==1.16.1
promptflow-core==1.16.1
promptflow-devkit==1.16.1
promptflow-rag==0.1.0
promptflow-tracing==1.16.1
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
psutil==5.9.8
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycomposefile==0.0.32
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydash==7.0.7
PyGithub==1.59.1
Pygments==2.18.0
PyJWT==2.9.0
pymsalruntime==0.17.1
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pypdf==3.17.4
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.20.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scp==0.13.6
SecretStorage==3.3.3
semver==2.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sshtunnel==0.1.5
starlette==0.41.2
strictyaml==1.7.3
tabulate==0.9.0
tenacity==8.5.0
tika==2.6.0
tiktoken==0.5.2
tqdm==4.66.6
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.4
unstructured-client==0.21.0
urllib3==2.2.3
waitress==2.1.2
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==3.1.1
wrapt==1.16.0
xmltodict==0.14.2
yarl==1.17.1
zipp==3.20.2
